fixing minimap to double

+++ src/data/utils/settings.c
@@ -1,7 +1,7 @@
 
 #include "data.h"
 
-void	set_position(int x, int y, double where[2])
+void set_position(double x, double y, double where[2])
 {
 	where[X] = x;
 	where[Y] = y;
